# app/docs/swagger.yaml
consumes:
- application/json
definitions:
  apperrors.ErrorCode:
    enum:
    - access_token_expired
    - all_signals_failed_processing
    - authentication_error
    - authorization_error
    - client_closed
    - database_error
    - forbidden
    - invalid_correlation_id
    - internal_error
    - invalid_request
    - invalid_url_param
    - malformed_body
    - not_implemented
    - network_error
    - password_too_short
    - refresh_token_invalid
    - request_too_large
    - unsupported_media_type
    - rate_limit_exceeded
    - resource_already_exists
    - resource_expired
    - resource_in_use
    - resource_not_found
    - timeout
    - token_creation_failed
    type: string
    x-enum-varnames:
    - ErrCodeAccessTokenExpired
    - ErrCodeAllSignalsFailedProcessing
    - ErrCodeAuthenticationFailure
    - ErrCodeAuthorizationFailure
    - ErrCodeClientClosed
    - ErrCodeDatabaseError
    - ErrCodeForbidden
    - ErrCodeInvalidCorrelationID
    - ErrCodeInternalError
    - ErrCodeInvalidRequest
    - ErrCodeInvalidURLParam
    - ErrCodeMalformedBody
    - ErrCodeNotImplemented
    - ErrCodeNetworkError
    - ErrCodePasswordTooShort
    - ErrCodeRefreshTokenInvalid
    - ErrCodeRequestTooLarge
    - ErrCodeUnsupportedMediaType
    - ErrCodeRateLimitExceeded
    - ErrCodeResourceAlreadyExists
    - ErrCodeResourceExpired
    - ErrCodeResourceInUse
    - ErrCodeResourceNotFound
    - ErrCodeTimeout
    - ErrCodeFailedToCreateToken
...
  handlers.WithdrawSignalRequest:
    properties:
      local_ref:
        example: item_id_#1
        type: string
    type: object
  responses.ErrorResponse:
    properties:
      error_code:
        allOf:
        - $ref: '#/definitions/apperrors.ErrorCode'
        description: Standard app error code
        example: error_code
      message:
        description: Client message
        example: client message
        type: string
      request_id:
        description: request tracking id
        example: a1b2c3d4
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Signals ISN service API for managing Information Sharing Networks

    ## Common Error Responses
    All endpoints may return:
    - `400` Malformed request (invalid json, missing required fields, etc.)
    - `401` Unauthorized (invalid credentials)
    - `403` Forbidden (insufficient permissions)
    - `404` Resource not found (e.g. an ISN or signal type that doesn't exist or is not in use)
    - `413` Request body exceeds size limit
    - `429` Rate limit exceeded
    - `500` Internal server error
    - `504` Request timed out (`timeout`)

    Individual endpoints document their specific business logic errors.

    Standard error responses include a JSON response body containing `error_code`, `message` and `request_id` fields.

    The /oauth endpoints return `error` (the RFC6749 standard error code), `error_code` and `error_description` fields instead.

    ## Request Limits
    The endpoints are protected by:
    - **Rate limiting**: Configurable requests per second (all endpoints)
    - **Request size limits** (configurable - these are the defaults): 64KB for admin/auth endpoints, 5MB for json and event signal submissions, 20MB for document uploads

    Check the `Signalsd-Max-Request-Size` response header for the configured size limit of the endpoint called.

    The rate limit is set globally and prevents abuse of the service.
...

    ## Authentication & Authorization

    ### OAuth
...

...

...
    ```
...
    ```

...
paths:
...
  /api/isn/{isn_slug}/signal-types/{signal_type_slug}/v{sem_ver}/signals/withdraw:
    put:
      description: |-
        Withdraw a signal by local reference

        Only the account's signal with this local reference on this ISN is withdrawn (a signal with the same local reference on another ISN is not affected).

        Withdrawn signals are hidden from search results by default but remain in the database.
        Signals can only be withdrawn by the account that created the signal.
        To reactivate a json signal, resubmit it with the same local_ref using the *Submit Signals* endpoint. To reactivate a document, upload it again with the same local_ref using *Upload a Document*.
        Withdrawn events cannot be reactivated - send a new event with a new local_ref instead.
      parameters:
      - description: ISN slug
        example: sample-isn
        in: path
        name: isn_slug
        required: true
        type: string
      - description: Signal type slug
        example: signal-type-1
        in: path
        name: signal_type_slug
        required: true
        type: string
      - description: version
        example: 1.0.0
        in: path
        name: sem_ver
        required: true
        type: string
      - description: Withdrawal request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.WithdrawSignalRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: malformed_body
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: authorization_error | access_token_expired
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: resource_not_found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: resource_already_exists
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "413":
          description: request_too_large
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: database_error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Withdraw a Signal
      tags:
      - Signal Exchange
...
produces:
...
securityDefinitions:
  BearerAccessToken:
    description: Bearer {JWT access token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...